chore(web): rename SiTextBox2 to SiTextBox
