+++ app/web/src/organisms/Secret/SecretCreateFields.vue
@@ -28,7 +28,7 @@
         >
       </div>
       <div class="w-1/2 align-middle">
-        <SiTextBox2
+        <SiTextBox
           :id="idFor(field.keyName, field.password)"
           v-model="secretMessage[field.keyName]"
           size="xs"
@@ -43,7 +43,7 @@
 
 <script setup lang="ts">
 import { ref } from "vue";
-import SiTextBox2 from "@/atoms/SiTextBox2.vue";
+import SiTextBox from "@/atoms/SiTextBox.vue";
 import { SecretKindFields } from "@/api/sdf/dal/secret";
 
 const emit = defineEmits(["update:modelValue"]);
