+++ app/web/src/organisms/deprecated/ResourceViewer.vue
@@ -34,7 +34,7 @@
             {{ new Date(resource.updatedAt) }}
           </div>
         </div>
-        <SiTextBox2
+        <SiTextBox
           id="resourceJson"
           name="resourceJson"
           :placeholder="JSON.stringify(resource)"
@@ -50,7 +50,7 @@
 import * as Rx from "rxjs";
 import { ref, toRefs, computed } from "vue";
 import { Resource, ResourceHealth } from "@/api/sdf/dal/resource";
-import SiTextBox2 from "@/atoms/SiTextBox2.vue";
+import SiTextBox from "@/atoms/SiTextBox.vue";
 import { ComponentService } from "@/service/component";
 import { GlobalErrorService } from "@/service/global_error";
 import { ChangeSetService } from "@/service/change_set";
